add get message order by date and add logout function

+++ lib/feature/home/presentation/screen/home_screen.dart
@@ -1,8 +1,10 @@
 import 'package:chat/core/ingect/get_it_control.dart';
+import 'package:chat/core/network/firebase/firebase_control.dart';
 import 'package:chat/core/network/firebase/message_model.dart';
 import 'package:chat/feature/home/presentation/logic/btn_cubit.dart';
 import 'package:chat/feature/home/presentation/logic/home_cubit.dart';
 import 'package:chat/feature/home/presentation/widget/app_bar_widget.dart';
+import 'package:chat/feature/home/presentation/widget/chat_friend_widget.dart';
 import 'package:chat/feature/home/presentation/widget/chat_send_widget.dart';
 import 'package:chat/feature/home/presentation/widget/text_form_widget.dart';
 import 'package:chat/feature/login/presentation/screen/export.dart';
@@ -34,12 +36,25 @@ class HomeScreen extends StatelessWidget {
                       List<MessageModel> list = state.listMessage ?? [];
                       return ListView.builder(
                         itemCount: list.length,
-                        itemBuilder: (context, index) => ChatSendWidget(
-                          title: list[index].textMsg,
-                          onDismissed: (){
-                            HomeCubit.get(context).deleteMessage(index);
-                          },
-                        ),
+                        itemBuilder: (context, index) => list[index].uid ==
+                                FireBaseControl.getUserId()
+                            ? ChatSendWidget(
+                                title: list[index].textMsg,
+                                onDismissed: () {
+                                  if (list[index].uid ==
+                                      FireBaseControl.getUserId()) {
+                                    HomeCubit.get(context).deleteMessage(index);
+                                  }
+                                },
+                              )
+                            : ChatFriendWidget(
+                                title: list[index].textMsg,
+                                onDismissed: () {
+                                  if (list[index].uid ==
+                                      FireBaseControl.getUserId()) {
+                                    HomeCubit.get(context).deleteMessage(index);
+                                  }                                },
+                              ),
                       );
                     }
                     return const SizedBox();
